Mapper task development

+++ app/src/main/java/org/closure/app/UserModule/dto/UserRequest.java
@@ -13,5 +13,5 @@ public class UserRequest {
     private String name;
     private String email;
     private String password;
-    private int community;
+    private Long community;
 }
